+++ test/reactor/ResolutionLib.Fuzz.t.sol
@@ -30,7 +30,7 @@ contract ResolutionLibFuzzTest is Test {
         co.order.input.amount = inAmount;
         co.order.output.amount = limit;
         co.order.output.maxAmount = maxOut;
-        co.order.slippage = slippage;
+        co.order.slippage = uint32(slippage);
         co.cosignatureData.input = OrderLib.CosignedValue({token: address(1), value: inputValue, decimals: 18});
         co.cosignatureData.output = OrderLib.CosignedValue({token: address(2), value: outputValue, decimals: 18});
